refactor: add new module replay_back
